pyproject.toml
[tool.poetry]
name = "ape-dasy"
version = "0.1.0"
description = ""
authors = ["z80 <z80@ophy.xyz>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
dasy = "0.1.18"
eth-ape = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = ">=7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
